chore: remove old editor components fix: fix preview error
