+++ app/(main)/(routes)/documents/[documentId]/page.tsx
@@ -17,7 +17,7 @@ interface DocumentIdPageProps {
 
 const DocumentIdPage = ({ params }: DocumentIdPageProps) => {
   const Editor = useMemo(
-    () => dynamic(() => import('@/components/editor_next'), { ssr: false }),
+    () => dynamic(() => import('@/components/editor'), { ssr: false }),
     []
   )
   const document = useQuery(api.documents.getById, {
